# gen/http/openapi.yaml
swagger: "2.0"
info:
    title: Programme.lv backend
    description: Service for managing users, tasks, and submissions.
    version: 0.0.1
host: localhost:80
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
    /auth/current/jwt:
        get:
            tags:
                - users
            summary: queryCurrentJWT users
            description: Query current JWT
            operationId: users#queryCurrentJWT
            parameters:
                - name: Authorization
                  in: header
                  description: JWT token used for authentication
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/UsersQueryCurrentJWTResponseBody'
                "400":
                    description: Bad Request response.
                    schema:
                        type: string
                "401":
                    description: Unauthorized response.
                    schema:
                        type: string
                "404":
                    description: Not Found response.
                    schema:
                        type: string
                "409":
                    description: Conflict response.
                    schema:
                        type: string
            schemes:
                - http
            security:
                - jwt_header_Authorization: []
...
    /users:
        get:
            tags:
                - users
            summary: listUsers users
            description: |-
                List all users

                **Required security scopes for jwt**:
                  * `users:read`
            operationId: users#listUsers
            parameters:
                - name: Authorization
                  in: query
                  description: JWT token used for authentication
                  required: false
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: array
                        items:
                            $ref: '#/definitions/UserResponse'
                "400":
                    description: Bad Request response.
                    schema:
                        type: string
                "401":
                    description: Unauthorized response.
                    schema:
                        type: string
                "404":
                    description: Not Found response.
                    schema:
                        type: string
                "409":
                    description: Conflict response.
                    schema:
                        type: string
            schemes:
                - http
            security:
                - jwt_query_Authorization: []
        post:
            tags:
                - users
            summary: createUser users
            description: Create a new user
            operationId: users#createUser
            parameters:
                - name: CreateUserRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/UsersCreateUserRequestBody'
                    required:
                        - username
                        - email
                        - firstname
                        - lastname
                        - password
            responses:
                "201":
                    description: Created response.
                    schema:
                        $ref: '#/definitions/UsersCreateUserResponseBody'
                        required:
                            - uuid
                            - username
                            - email
                            - firstname
                            - lastname
                "400":
                    description: Bad Request response.
                    schema:
                        type: string
                "401":
                    description: Unauthorized response.
                    schema:
                        type: string
                "404":
                    description: Not Found response.
                    schema:
                        type: string
                "409":
                    description: Conflict response.
                    schema:
                        type: string
            schemes:
                - http
    /users/{uuid}:
        get:
            tags:
                - users
            summary: getUser users
            description: |-
                Get a user by UUID

                **Required security scopes for jwt**:
                  * `users:read`
            operationId: users#getUser
            parameters:
                - name: Authorization
                  in: query
                  description: JWT token used for authentication
                  required: true
                  type: string
                - name: uuid
                  in: path
                  description: UUID of the user
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/UsersGetUserResponseBody'
                        required:
                            - uuid
                            - username
                            - email
                            - firstname
                            - lastname
                "400":
                    description: Bad Request response.
                    schema:
                        type: string
                "401":
                    description: Unauthorized response.
                    schema:
                        type: string
                "404":
                    description: Not Found response.
                    schema:
                        type: string
                "409":
                    description: Conflict response.
                    schema:
                        type: string
            schemes:
                - http
            security:
                - jwt_query_Authorization: []
        delete:
            tags:
                - users
            summary: deleteUser users
            description: |-
                Delete a user

                **Required security scopes for jwt**:
                  * `users:write`
            operationId: users#deleteUser
            parameters:
                - name: Authorization
                  in: query
                  description: JWT token used for authentication
                  required: true
                  type: string
                - name: uuid
                  in: path
                  description: UUID of the user
                  required: true
                  type: string
            responses:
                "204":
                    description: No Content response.
                "400":
                    description: Bad Request response.
                    schema:
                        type: string
                "401":
                    description: Unauthorized response.
                    schema:
                        type: string
                "404":
                    description: Not Found response.
                    schema:
                        type: string
                "409":
                    description: Conflict response.
                    schema:
                        type: string
            schemes:
                - http
            security:
                - jwt_query_Authorization: []
definitions:
...
    UserResponse:
        title: UserResponse
        type: object
        properties:
            email:
                type: string
                description: Email of the user
                example: johndoe@example.com
                format: email
            firstname:
                type: string
                description: First name of the user
                example: John
            lastname:
                type: string
                description: Last name of the user
                example: Doe
            username:
                type: string
                description: Username of the user
                example: johndoe
            uuid:
                type: string
                description: Unique user UUID
                example: 550e8400-e29b-41d4-a716-446655440000
        description: User representation
        example:
            email: johndoe@example.com
            firstname: John
            lastname: Doe
            username: johndoe
            uuid: 550e8400-e29b-41d4-a716-446655440000
        required:
            - uuid
            - username
            - email
            - firstname
            - lastname
    UsersCreateUserRequestBody:
        title: UsersCreateUserRequestBody
        type: object
        properties:
            email:
                type: string
                description: Email of the user
                example: johndoe@example.com
                format: email
            firstname:
                type: string
                description: First name of the user
                example: John
            lastname:
                type: string
                description: Last name of the user
                example: Doe
            password:
                type: string
                description: Password of the user
                example: password123
            username:
                type: string
                description: Username of the user
                example: johndoe
        example:
            email: johndoe@example.com
            firstname: John
            lastname: Doe
            password: password123
            username: johndoe
        required:
            - username
            - email
            - firstname
            - lastname
            - password
    UsersCreateUserResponseBody:
        title: UsersCreateUserResponseBody
        type: object
        properties:
            email:
                type: string
                description: Email of the user
                example: johndoe@example.com
                format: email
            firstname:
                type: string
                description: First name of the user
                example: John
            lastname:
                type: string
                description: Last name of the user
                example: Doe
            username:
                type: string
                description: Username of the user
                example: johndoe
            uuid:
                type: string
                description: Unique user UUID
                example: 550e8400-e29b-41d4-a716-446655440000
        example:
            email: johndoe@example.com
            firstname: John
            lastname: Doe
            username: johndoe
            uuid: 550e8400-e29b-41d4-a716-446655440000
        required:
            - uuid
            - username
            - email
            - firstname
            - lastname
    UsersGetUserResponseBody:
        title: UsersGetUserResponseBody
        type: object
        properties:
            email:
                type: string
                description: Email of the user
                example: johndoe@example.com
                format: email
            firstname:
                type: string
                description: First name of the user
                example: John
            lastname:
                type: string
                description: Last name of the user
                example: Doe
            username:
                type: string
                description: Username of the user
                example: johndoe
            uuid:
                type: string
                description: Unique user UUID
                example: 550e8400-e29b-41d4-a716-446655440000
        example:
            email: johndoe@example.com
            firstname: John
            lastname: Doe
            username: johndoe
            uuid: 550e8400-e29b-41d4-a716-446655440000
        required:
            - uuid
            - username
            - email
            - firstname
            - lastname
...
    UsersQueryCurrentJWTResponseBody:
        title: UsersQueryCurrentJWTResponseBody
        type: object
        properties:
            audience:
                type: array
                items:
                    type: string
                    example: Modi minima et illo dolor.
                example:
                    - Quasi accusamus.
                    - Repellendus et molestias.
            email:
                type: string
                example: Tenetur temporibus harum assumenda.
            expires_at:
                type: string
                example: Eum quia.
            firstname:
                type: string
                example: Quam est repudiandae.
            issued_at:
                type: string
                example: Et nulla.
            issuer:
                type: string
                example: Qui repellat.
            lastname:
                type: string
                example: Quo at.
            not_before:
                type: string
                example: Quisquam laboriosam optio qui.
            scopes:
                type: array
                items:
                    type: string
                    example: Ipsa exercitationem.
                example:
                    - Voluptatem et nobis exercitationem.
                    - Consequatur voluptas neque tenetur nemo cumque.
            subject:
                type: string
                example: Voluptatibus quae autem odio.
            username:
                type: string
                example: Et doloremque.
            uuid:
                type: string
                example: Harum laboriosam est hic.
        example:
            audience:
                - Explicabo occaecati aperiam.
                - Et consectetur iure quibusdam.
                - Molestias explicabo odio quasi est.
                - Quos et occaecati rem eius a.
            email: Quis repellat animi soluta nulla aut dolorum.
            expires_at: Totam praesentium dolor aspernatur doloremque est et.
            firstname: Saepe eos debitis ratione tempora.
            issued_at: Eveniet omnis aut a nihil blanditiis deserunt.
            issuer: Vitae voluptate eum est.
            lastname: Vel voluptatem est accusamus animi et.
            not_before: Laborum quia repellat voluptatem quia voluptas.
            scopes:
                - Est sit.
                - Reiciendis incidunt ratione.
                - Fugiat aut sint reiciendis.
            subject: Quod est laudantium et aliquid quo magni.
            username: Sint nostrum.
            uuid: Repudiandae accusamus voluptas.
securityDefinitions:
    jwt_header_Authorization:
        type: apiKey
        description: |4-
            **Security Scopes**:
              * `users:read`: Read users
              * `users:write`: Write users
        name: Authorization
        in: header
    jwt_query_Authorization:
        type: apiKey
        description: |4-
            **Security Scopes**:
              * `users:read`: Read users
              * `users:write`: Write users
        name: Authorization
        in: query